Basic support for strings, keywords, and operators.
Basic support for strings; interpolated strings are not treated any
differently.
Basic support for keywords is in. However, there may be a bug in the VS
Code regex handling that [^`] is not valid for properly ignore the
`class` keyword/identifier case.
Partial support for a few of the context sensitive keywords.
Basic operator support; no unicode support.